refactor: move list group
